Print the project name while logging
Especially in long builds, it is always tricky to find the portion
corresponding to a specific project in the log. This commit adds
the project name to the log messages, making it much easier to
scan the log for information. Note: this may possibly break
those scripts that manually parse the log messages.