Change the generation of the "repositories" file in SbtBuildRunner
The previous idea was: generate a "repositories" file containing
just the resolvers from the config file, then add at each level
the repository with the rematerialized artifacts, using onLoad.
For the top level (where no rematerialized artifacts exist),
only use the defined resolvers.

That technique worked great, *but* failed with sbt itself:
while grabbing the Scala compiler and library, sbt ignores
the rewritten fullResolvers setting, fetching stuff instead
directly from the resolvers in the "repositories" file only.

That leads to an interesting problem: the list used to grab
Scala compiler and library, and the list used at the top level
will always be the same. Also, the Scala compiler and library
will always be grabbed from the same places, for ALL levels
(including plugins).

That slightly defeats our attempt at isolating the levels,
using a different space for each. We have to resort to the
following workaround:
1) we add the repo for the rematerialized artifacts for the
base level to the "repositories" file
2) from onLoad, we strip that repo and add instead the right
one for the level we are in currently. Which works for everything,
except grabbing the scala compiler and library in sbt.

Consequence: the top level will find on its "repositories"
list also the rematerialized artifacts of the ground level.
Although that should not be much of a problem in practice,
in theory the rematerialized artifacts could interfere
with the default one, which is why isolation is not perfect
in this case; that is due to the above mentioned sbt limitation.