Respect the flags --no-resolvers and --local for "dbuild checkout"
If the saved repeatable configuration contains local resolvers,
it was already possible to override them in dbuild by using the
flag --no-resolvers, which causes dbuild to use only the
repositories defined in dbuild.properties, instead.

This commit enables the same behavior for "dbuild checkout":
the user can use --no-resolvers to override the list of
repositories saved in the repeatable build for the given
uuid.