Update extraction-version to work with onLoad-based extraction
Prior to dbuild 0.9, extraction-version was implemented using
an additional "set" command. Commands are now not executed at
all during extraction (nor any other command is run, as all
works directly within onLoad); therefore, it is necessary to
set the extraction Scala version using a "fixer" setting similar
to those used in DistributedRunner, which in turn requires us
to save and restore the previous onLoad handler; if that is not
done, newState() invokes again onLoad, resulting in an endless
loop.