Some initial work towards improving the cycle detection messages.
Right now, if a cycle is detected the resulting message is extremely
long and confusing, and not very useful: all the projects in each
strongly connected component are printed, with their list of
dependencies, meaning that one then has to search painstakingly
among hundreds of dependencies.

This commit adds a witness to the build graph edges; it also preserves
the ordering of nodes found while looking for cycles, so that a
better message can be eventually printed.