Small but crucial fix to serialization/deserialization of settings
The default value for settings (which are a SeqSeqString) used to
be SeqSeqString(Seq.empty), meaning an empty list of per-level
settings (no levels). That was ok, except that the value is
serialized as "[]". Upon reloading, the SeqSeqStringDeserializer
would turn that into a SeqSeqString containing instead *one*
element: an empty list of settings (for the main build level).

Although the net effect of the two values in the code is identical,
the difference would cause repository consistency check failures,
causing builds to abort when repeated identical.

Using a different default value results in the same value being
written and later read, solving this issue.