Generating ModuleInfo in all build systems; Assemble needs changes
At this time, the special hackish "assemble" build system compiles
each of the componend projects, and then assembles them into a
fictitious project that has each of the original component projects
as its own subprojects.

Which worked great until now. However, now the module/subproject
alignment is shattered, since the artifacts produced by one
subproject corresponds not to a single one, but to a bunch of
modules. So, either I change BuildArtifactsOut so that possibly
a list of ModuleInfo is returned (with an unknown relationship
between modules and artifacts), or I chance assembly so that
the original subprojects are preserved (but then the names of
the subprojects need to be changed, in order to avoid duplications).