Better calculation of subprojects in Assemble
There was still a misalignment of the recalculated subproject
names between extracted data and final building data. That was
not a problem as long as no subproject selection was performed
on the assembled subprojects, but a partial deploy might have
failed otherwise.
The recalculation of subprojects is now refactored, and common
between extraction and building in Assemble.