Fixed some issues with distribution, and root aggregate
The root aggregate was not including the "deploy" project;
now fixed.

Also: the "dist" project creates "dbuild" artifacts
containing the launcher, in zip and tgz format. That
would collide with the "dbuild" subproject. Now the
"dbuild" and "drepo" subprojects have been renamed
to "build" and "repo" respectively, in order to
avoid the collision.