Fix issue where sometimes the user configured more build-levels than we extracted.
It can occur that we have not extracted any information (indeed, sbt never even loaded)
a given level because the user has configured injected settings for a build level that
is otherwise not used.   When this happens, we should still generate all
necessary input files, but we can return blank/empty artifacts for these
levels as we will generally ignore them.