If there are injected settings, consider them while calculating levels
In a dbuild configuration file, there might be a "settings" option that
inject additional settings on more levels than there actually are in
the source code of a given sbt-based project. In that case, the total
number of levels must be calculated including the additional ones
supplied by "settings".

Further, both during extraction and building the calculated number
of levels must be identical: extraction will calculate the set of
artifacts which will eventually be included in the RewireInfo that
will be used to rewire each level.

This commit also simplifies the calculation of levels as modified
by 4189f9e31e3011caad0cbda28ac68ebe0f5db1cd. In particular:
- the subprojs sequence is automatically calculated during
  extraction, and is the same length of the number of levels
  calculated during that stage
- crossVers is only meaningful for the levels that actually
  generate artifacts, and the same applies to checkMissing;
  consequently the excess ones can simply be ignored.

The calculation can be therefore simplified to:
SbtRunner.buildLevels(projectDir) max sbtSettings.size