Implementation of custom list of test tasks
This commit adds the new sbt-specific option "test-tasks", which
can be used to customize the testing stage.

It also adds the new "post-commands" and "sbt-post-commands",
which are the counterparts of the existing "commands" and
"sbt-commands". The post-commands are run after building and
testing, and can also be used to run arbitrary shell commands
after testing has completed.