Temporarily moving out of the way the incomplete Maven build system
The current (incomplete) code is using an outdated version of Aether,
resulting in some bitrot. The package names would have to be changed,
and the logic reworked in order to conform to the new API and package
names.

For the moment, this incomplete build system is being moved out of
the way, so that the new mini-build system based on Aether can use
the latest API; at a later time, the old code should be brought up
to date again.