Remove "standard" cross-version from aether and assemble.
The "standard" cross-versioning level is defined in a precise
manner only for the "sbt" build system, in which it means:
"pick what sbt would normally select depending on the user's
choice of scalaBinaryVersion".

For the other build systems we may work out an approximation
of the same behavior, but it is in general more clear to
directly select one of "disabled", "binary", "full" instead.