Correctly support "binary" cross-version in the Scala build system
The property "maven.version.number" is used to describe the full
scala version number of the resulting artifacts; however, a separate
property scala.binary.version is used to create the cross-version
suffixes in the pom files.

As long as the compiler was generated inside an "Assemble" section,
that would not make a difference, since the suffix was stripped
away and replaced while assembilng, but in a bootstrap process that
requires a binary cross-version, that is important.