Added docs clarification.
I ran a few tests: the generic bintray upload syntax ("/content")
does actually work for maven-style repositories as well, even if
the package version number is not the same as the artifacts that
are being uploaded (which makes sense considering that perfectly
regular Maven artifacts with different version numbers may be
part of a single "release", for instance one module with one
version and another one with a different version number).

In any case, the bintray deploy uri format does seem to work
for generic as well as for maver-style bintray repos.