Implementing "standard" cross-versioning for Aether and Assembly
It is now possible to use "standard" as a possible cross-versioning
level for both the Aether and the Assemble build system. For
Aether, using "standard" means that the resulting cross-versioning
suffix during deploy will be the same one present on the original
artifact (as opposed to the already existing "disabled", "full",
and "binary", which always force a specific format).

In the case of the Assemble build system, "standard" will mean
that the format for each individual artifact in the assembly
is preserved, but the Scala base version used to generate the
suffix will be taken from the Scala library, if one is present
within the assembly. For instance, let us assume that the Scala
library is a hypothetical version "2.13.94". A suffix like
"_2.10" will become "_2.13", while "_2.10.4" will become
"_2.13.94". Each module in the assembly can have a different
suffix. If no Scala library is found, the existing cross-version
suffixes are preserved as-is.