Some tricky magic to include bintray both in 0.13 and 1.0.0-M4
If we are on sbt 0.13, then a regular bintray plugin dependency
will be included (but only on Scala 2.10) by the file:
project/plugins-2.10.sbt.
When this file is seen while we use 1.0.0-M4, libraryDependencies
is not modified and no Bintray plugin is inserted.

If we are on sbt 1.0.0-M4, an additional project will be creaated
by the file:
project/project/src/main/scala-2.11/bintrayIf211.scala
However, when we are on 0.13, this file is used instead:
project/project/src/main/scala-2.10/bintrayIf211.scala
The definition consists in a Seq of projects (empty for 2.10),
which is then added by a dependsOn clause in project/build.sbt