Turn sbt dependencies into "provided" ones
The dbuild plugins used to bring onto the classpath all sorts of
sbt libraries; the result was that multiple versions of the sbt
libraries would appear on the classpath, leading to bizarre
behaviour. The plugin will now use whatever sbt places on the
classpath, matching the sbt version in use.

In theory, this should fix scala/community-builds#384 and may
help with sbt/sbt#1696