Revert to dispatch classic to grab sbt-launch.jar
The dispatch reboot, at least in v0.11.3 (the last
available on Scala 2.10, needed by sbt 0.13), seems
to have a bug related to URL handling, by which a
query string that contains multiple '=' signs within
a single query component gets mangled. That causes
the URL to get corrupted, and downloads from the
Akamai frontend of Bintray to fail.

This bug will probably affect the dispatch used in
the regular code, so specific testing is needed,
and a request for a patch submitted to the 'dispatch'
library if needed.