Change to dbuild plugin detection logic
The logic condition in DBuildRunner was incorrect, as it was
excluding from the list of "missing" plugins, in case the full
tracking of dependencies is requested, both all the plugins
in the "com.typesafe.dbuild", as well as those named "plugin"
(a logical or, rather than an and). Using just the organization
is safer, in any case, so the condition has been changed slightly.