Removed the 2.12/2.10 specific dependency build files
This is required for cross-compilation compatibility: the
cross-building infrastructure will only compile at startup
the scala version-specific build source files that correspond
to the default scala version in use, but will not switch when
a different scala version is selected. So, we must get rid of
the `project/src/main/scala-2.1?/xxx.scala` files.