Timeouts work as intended (but underlying processes keep running)
The timeout handling now works as it should: in case extraction
or build take longer than the allowed time, a timeout outcome
will be returned to the main.

This will fix the main issue; however, there are still two
pending problems:
- In case of a timeout, an underlying Process (sbt, for example)
will keep running until completion (or forever), even after
dbuild and the whole jvm it runs in terminate. That will wreak
havoc in case of an endless test, for instance, which will
burn cpu indefinetely, or if a subsequent build is attempted
in the same dir, which will collide with the previous run.
- A minor issue is that the logger actors do not seem to
terminate properly as they should. They will go away in any
case once the jvm exits.