Even better timeouts for build phase
Rather than failing every build, preserve the status of those
already completed in the final report. That is done by adding,
rather than a single watchdog on the mapped sequence of futures,
a separate watchdog for each build, so that the result of each
will be the first completed of the regular build, or the watchdog.
The final status is then collected.