Consider 'skip in publish' while extracting and building artifacts
The `skip in publish` flag must be considered when getting the
list of dependencies artifacts, during extraction and
building. If skip in publish is set, the list of dependencies
is kept, but we consider that no artifact is produced.

This is only partly correct, since the generated artifacts would still be visible by
the inter-project resolver, even if they are not published. On the other hand, builds
currently performed by dbuild, prior to this patch, would also fail to publish
artifacts that can be seen by other subprojects since "publish" is hijacked to collect
artifacts.

Fixes #205