Optimizations for the build graph
The BuildGraph calculation was performed twice, once after extraction,
and one in case of a repeated build when deserializing the previously
serialized RepeatableDBuildConfig, to make sure it matches.

That was not supposed to be the case: the "graph" value inside the
RepeatableDBuildConfig is actually a lazy val. However, an inappropriately
placed "checkCycles()" would always trigger the graph creation. This
commit moves checkCycles() to the end of the BuildGraph creation,
so that the graph is not immediately built upon deserialization.

Also, this commit makes "topological" in graph a lazy val, for a further
improvement. Note that "checkCycles()" already relied on "isCyclic",
which is also a lazy val.