Split the extractor resolve() into resolve/prepare
We will have resolve() determine the stable URI (for instance,
converting a branch/tag into a commit hash), while prepare()
will be in charge of preparing the source tree.

The existing implementation of ProjectResolver currently do
all the work in resolve(), while prepare() is empty; however,
we will shortly convert the git resolver so that reference
resolution and source tree setup are distinct. The reason is
that in many cases an extracted dependency set will already be
in the cache: we want to be able to look into the cache, but
there is no point in preparing a source tree if we are not
going to rebuild/reextract.