Allow for immediate deletion of working data
Normally dbuild will automatically reclaim the space
used by working dirs where projects are built or
extracted, after a configurable age. The reclaiming
takes place at the start of each dbuild run.

If the configurable age is set to zero, all of the
old data will be deleted at the beginning of the run;
the data corresponding to the current run will
however remain available at the end.

This commit allows the age to be set to negative
values. If that is the case, the temporary working
data will be deleted immediately after each extraction
or building of each project, saving disk space. It
is still possible, for instance, to specify that
we want the temporary data of the projects that
succeeded to be discarded immediately, while the
data of the failed projects can be kept in order
to diagnose the failure.