Optimization: look if the hash is already before fetching
If the ref that we are resolving looks like a commit hash,
then look if it is already present in the cached clone before
performing a full fetch, which would unshallow the central clone.

That also prevents the build stage (in which all refs are hashes)
from unshallowing all of the clones as it progresses.

A possible further improvement would be heuristically deciding that
refs that look like hashes and are exactly 40 characters long most
probably are indeed hashes, therefore skip the lookup for a
matching branch or tag.