Treat references that look like a 40-chars shas as commit hashes
We heuristically decide that
refs that look like hashes and are exactly 40 characters long most
probably are indeed hashes, therefore skip the lookup for a
matching branch or tag.

We could possibly check, in case that didn't work, if it is a
bizarre 40-character branch name or tag name that looks exactly
like a hash, but that is improbable enough; at present we just
abort.