Don't checkMissing when fixing dependencyOverrides
When fixing overrides, I see very little need to bomb with
"checkMissing".  You can still emit all the warnings and everything, but
just don't bomb.  If the dependency is also in allDependencies, then
it'll still bomb.

This is relevant because there's a dependencyOverrides on
scripted-plugin since sbt 1.2, which blows up in fixModule.